Amplify cosmos tx priority (#443)
## Describe your changes and provide context
Since EVM transactions specify gas price in unit of wei, we have started
representing priority of EVM transactions in wei-per-gas. In order for
cosmos transactions' priority to be comparable with EVM transactions',
we would like to use the same unit for priority here as well. Hence in
this PR we amplify cosmos transaction priority by 10^12 (if the original
priority is based on the base denom aka usei)

## Testing performed to validate your change
unit test