Refactor readset to handle potential race conditions (#449)
## Describe your changes and provide context
In some rare situations, it is possible that the race between iterators
and other potential reads interleaved with writes to the same key by
earlier transactions may result in stale values being used by a
transaction but not being caught during validation. This refactor of
readset behavior rigorously requires that ALL attempts to update the
readset must either ADD to the readset if the value is different from
that previously encountered, or no-op only if that value is already
existing in the readset. As a result, if we encounter a situation where
a race condition caused different values to be read for a key, this will
catch it in validation and result in the transaction failing validation
and requiring re-execution.

## Testing performed to validate your change
Unit tests