Adding a test to ensure that tx commit order is properly tested
Tracers can expect that that `OnTxCommit` is called in proper transaction order so that it have the certitude that the current `OnTxCommit` is that correct one to happen relative to all other transaction.

This is already the case, this test case only add a certain confidence that this assertion is going to be respected if any refactor happen in the future.