Add whitelist for fee denoms (#564)
## Describe your changes and provide context
It is possible for tx sender to include "garbage" tokens as fees which
may decrease chain performance if in large quantity (in terms of the
number of different denoms, which may be created via tokenfactory). This
PR introduces a new parameter to whitelist denoms that can be used for
fees. Note that `usei` is always whitelisted.

## Testing performed to validate your change
local sei with tokenfactory tokens